[ui] Catch cookie.set exceptions in supabase server client
This addresses the following issue:
https://github.com/vercel/next.js/issues/51875

The changes were inspired by the create-next-app template "with-supabase":

https://github.com/vercel/next.js/blob/canary/examples/with-supabase/utils/supabase/server.ts

Also, for readability split supabase clients in:
- lib/supabase/server
- lib/supabase/middleware
