+++ app/src/app/_actions/resetLegalUnits.ts
@@ -1,6 +1,7 @@
 "use server";
 
-import {createClient} from "@/lib/supabase.server.client";
+
+import {createClient} from "@/lib/supabase/server";
 
 export async function resetLegalUnits() {
   "use server";
