+++ app/src/app/_actions/resetRegions.ts
@@ -1,6 +1,7 @@
 "use server";
 
-import {createClient} from "@/lib/supabase.server.client";
+
+import {createClient} from "@/lib/supabase/server";
 
 export async function resetRegions() {
     "use server";
