+++ app/src/app/_actions/resetSettings.ts
@@ -1,6 +1,7 @@
 "use server";
 
-import {createClient} from "@/lib/supabase.server.client";
+
+import {createClient} from "@/lib/supabase/server";
 
 export async function resetSettings() {
     "use server";
