+++ app/src/app/login/actions.ts
@@ -1,6 +1,6 @@
 'use server'
-import {createClient} from "@/lib/supabase.server.client";
 import {redirect} from "next/navigation";
+import {createClient} from "@/lib/supabase/server";
 
 export async function login(formData: FormData) {
   const email = String(formData.get('email'))
